Change on saves and private posts

+++ database/migrations/2022_03_23_183917_create_posts_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('image')->nullable();
             $table->string('video')->nullable();
             $table->longText('content');
-            $table->double('private')->default(0)->nullable();
+            $table->integer('private')->default(0)->nullable();
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
